+++ components/NothingHere/NothingHere.js
@@ -1,4 +1,5 @@
-import { Main, StyledLink } from "../../components/StyledNothingHere";
+import Link from "next/link";
+import styled from "styled-components";
 import { StyledHeading } from "../../pages/[category]";
 import Lottie from "lottie-react";
 import astronautMusic from "../../public/Lottie/astronautMusic.json";
@@ -22,3 +23,17 @@ export default function NothingHere() {
     </>
   );
 }
+
+export const Main = styled.section`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+export const StyledLink = styled(Link)`
+  text-decoration: none;
+  background-color: darkmagenta;
+  color: white;
+  padding: 0.5rem;
+  border-radius: 5px;
+`;
